chore: refactor API routes and update styling
- Removed old API routes and replaced them with new route handlers
- Updated email sending functionality to use Next.js API routes
- Migrated from nodemailer to SendGrid for email handling
- Refactored global CSS to use Tailwind CSS with custom theming
- Deleted unused PostCSS and Tailwind configuration files
- Added new PostCSS configuration in ES module format
- Updated TypeScript configuration for JSX support
- Introduced pnpm workspace configuration for dependency management
