+++ AGENT.md
@@ -4,46 +4,58 @@ A personal "coming soon" / pre-launch portfolio page for Gilead Odo, displayed w
 
 ## Tech Stack
 
-- **Framework**: Next.js 14 (Pages Router)
-- **Language**: TypeScript
-- **UI Library**: Mantine v7 (`@mantine/core`, `@mantine/hooks`, `@mantine/notifications`)
-- **Icons**: Tabler Icons (`@tabler/icons-react`)
-- **Styling**: Tailwind CSS v3
-- **Email**: SendGrid (`@sendgrid/mail`)
+- **Framework**: Next.js 16 (App Router)
+- **Language**: TypeScript 5.9
+- **UI Library**: shadcn/ui (`class-variance-authority`, `clsx`, `tailwind-merge`)
+- **Icons**: Lucide React (`lucide-react`)
+- **Styling**: Tailwind CSS v4 CSS-first (`tw-animate-css`, `shadcn/tailwind.css`)
+- **Theming**: next-themes 0.4 (with custom `ThemeProvider` + `ThemeHotkey`)
+- **Email**: SendGrid (`@sendgrid/mail`) + Nodemailer (alternative)
 - **HTTP Client**: Axios
-- **Package Manager**: pnpm (preferred), npm/yarn also present
+- **Package Manager**: pnpm
 
 ## Project Structure
 
 ```
 src/
+  app/
+    layout.tsx           # Root layout (Inter font, ThemeProvider)
+    page.tsx             # Home / landing page
+    globals.css          # Tailwind v4 CSS-first config (@theme tokens, CSS variables, base styles)
+    contact/
+      page.tsx           # Contact form page
+  components/
+    theme-hotkey.tsx     # Keyboard shortcut for theme toggling
+    theme-provider.tsx   # next-themes provider wrapper
+    ui/
+      button.tsx         # shadcn/ui Button
+      card.tsx           # shadcn/ui Card
+      input.tsx          # shadcn/ui Input
+      label.tsx          # shadcn/ui Label
+      textarea.tsx       # shadcn/ui Textarea
+  lib/
+    utils.ts             # cn() utility (clsx + tailwind-merge)
   pages/
-    index.tsx          # Home / landing page
-    contact.tsx        # Contact form page
-    _app.tsx           # App wrapper (Mantine provider setup)
-    _document.tsx      # Custom document
     api/
-      send-email.ts    # SendGrid email API route (active)
-      email.ts         # Nodemailer email API route (alternative)
-      hello.ts         # Default Next.js hello route
-  hooks/
-    useNotification.tsx  # Mantine notification helper hook
-  styles/
-    globals.css        # Global styles
-public/                # Static assets
+      send-email.ts      # SendGrid email API route (active)
+      email.ts           # Nodemailer email API route (alternative)
+      hello.ts           # Default Next.js hello route
+public/                  # Static assets
 ```
 
 ## Key Conventions
 
-- Pages use the Next.js Pages Router (not App Router)
-- Form validation is handled client-side in `contact.tsx` before submission
+- Pages use the Next.js App Router (`src/app/`)
+- API routes remain in `src/pages/api/` (Pages Router API coexisting with App Router)
+- shadcn/ui components live in `src/components/ui/`; shared components in `src/components/`
+- `cn()` from `src/lib/utils.ts` is used for conditional class merging
 - API responses follow the shape `{ entity?: string; message: string }`
-- Mantine components are used for UI elements; Tailwind is used for layout/spacing
+- Tailwind v4 CSS-first: no `tailwind.config.ts`; design tokens (colors, radii) are defined via `@theme inline` in `globals.css`; PostCSS wired via `postcss.config.mjs`; dark mode via `@custom-variant dark (&:is(.dark *))`
 
 ## Environment Variables
 
 | Variable           | Description                        |
-|--------------------|------------------------------------|
+| ------------------ | ---------------------------------- |
 | `SENDGRID_API_KEY` | SendGrid API key for sending email |
 
 Set these in a `.env.local` file (not committed to git).
